+++ src/utest/dispatcher.cpp
@@ -11,7 +11,7 @@ enum { CSTACK_SZ = 8192 };
 
 static void* myalloc(void*, void* old, aint_t sz)
 {
-    return realloc(old, sz);
+    return realloc(old, (size_t)sz);
 }
 
 static void add_test_module(aasm_t* a)
