+++ src/utest/gc.cpp
@@ -12,7 +12,7 @@ enum { CSTACK_SZ = 16384 };
 
 static void* myalloc(void*, void* old, aint_t sz)
 {
-    return realloc(old, sz);
+    return realloc(old, (size_t)sz);
 }
 
 static bool search_for(agc_t* gc, aint_t i)
