+++ src/utest/msbox.cpp
@@ -10,7 +10,7 @@ enum { CSTACK_SZ = 16384 };
 
 static void* myalloc(void*, void* old, aint_t sz)
 {
-    return realloc(old, sz);
+    return realloc(old, (size_t)sz);
 }
 
 static bool done;
